==> scripts/99_make_taxonomy.R <==
# Make taxonomy table by referencing `locus_taxonomy` file on Dropbox**

# dt <- read.csv("data/00_bathy_depth/00_bathy_depth.csv",
#                stringsAsFactors = FALSE)
# dt2 <- read.csv("data/depth_predictors.csv",
#                stringsAsFactors = FALSE)
# cat(unique(c(names(dt), names(dt2))))

suppressMessages(library(dplyr))

taxonomy_key <- bind_rows(
  c(col_name = "llid",
    description = "unique lake identifier developed for LAGOS-US"),
  c(col_name = "state",
    description = "abbreviation of the state used to search for a lake's depth",
    pretty = "State"),
  c(col_name = "max_depth_m",
    description = "lake maximum depth in meters",
    pretty = "Max depth (m)"),
  c(col_name = "mean_depth_m",
    description = "lake mean depth in meters",
    pretty = "Mean depth (m)"),
  c(col_name = "source",
    description = "url link to the source of depth data"),
  c(col_name = "lake_waterarea_ha",
    description = "surface area of lake waterbody polygon from NHD (excludes islands) by way of the LAGOS-US Locus module",
    pretty = "Area (ha)"),
  c(col_name = "lake_connectivity_permanent",
    description = "connectivity of focal lake to upstream features (DrainageLK = drainage lake with an upstream lake, Drainage = drainage lake with upstream stream, Headwater = lake with outlet but no inlet, Isolated = lake with no inlets or outlets) by way of the LAGOS-US Locus module"),
  c(col_name = "lake_lat_decdeg", description = "the latitude of the lake center point (NAD83) from the LAGOS-US Locus module"),
  c(col_name = "lake_lon_decdeg",
    description = "the longitude of the lake center point (NAD83) from the LAGOS-US Locus module"),
  c(col_name = "lat",
    description = "the latitude of the lake center point"),
  c(col_name = "long",
    description = "the longitude of the lake center point"),
  c(col_name = "lagoslakeid",
    description = "unique lake identifier developed for LAGOS-US"),
  c(col_name = "lake_maxdepth_m",
    description = "lake maximum depth in meters",
    pretty = "Max depth (m)"),
  c(col_name = "lake_meandepth_m",
    description = "lake mean depth in meters",
    pretty = "Mean depth (m)"),
  c(col_name = "maxdepth_true_true",
    description = "lake maximum depth calculated from inlake_slope and dist_deepest in meters"),
  c(col_name = "maxdepth_true_false",
    description = "lake maximum depth calculated from inlake_slope and dist_viscenter"),
  c(col_name = "maxdepth_false_true",
    description = "lake maximum depth calculated from slope_mean and dist_deepest in meters"),
  c(col_name = "maxdepth_false_false",
    description = "lake maximum depth calculated from slope_mean and dist_viscenter in meters"),

  c(col_name = "inlake_slope_pnt",
    description = "slope of the lake bottom assuming a straight line from the shore to the deepest point of the lake in units of depth change in meters per meter distance",
    pretty = "Inlake slope (m/m)"),

  c(col_name = "inlake_slope_pnts",
    description = "slope of the lake bottom assuming a straight line from the shore to the deepest point(s) of the lake in units of depth change in meters per meter distance",
    pretty = "Inlake slopes (m/m)"),

  c(col_name = "inlake_slope_mean",
    description = "the mean slope of the lake bottom in units of depth change in meters per meter distance",
    pretty = "Mean inlake slope (m/m)"),

  c(col_name = "inlake_slope_online_mean",
    description = "average slope of a line extending from the deepest point of a lake to the shoreline",
    pretty = "Inlake slope online (m/m)"),

  c(col_name = "inlake_slopes_online_mean",
    description = "average slope of a line (or lines) extending from the deepest point(s) of a lake to the shoreline",
    pretty = "Inlake slopes online (m/m)"),

  c(col_name = "nearshore_slope_mean",
    description = "average slope of the land surrounding a lake in a 100m buffer",
    pretty = "Nearshore mean slope (m/m)"),
  c(col_name = "buffer_slope",
    description = "average slope of the land surrounding a lake in a 100m buffer",
    pretty = "Nearshore slope (m/m)"),

  c(col_name = "nearshore_slope_online_mean",
    description = "average slope of the land surrounding a lake in a straight-line  extending from the deepest point of a lake and buffered on either side by 20m",
    pretty = "Nearshore slope online (m/m)"),

  c(col_name = "nearshore_slopes_online_mean",
    description = "average slope of the land surrounding a lake in a straight-line  extending from the deepest point(s) of a lake and buffered on either side by 20m",
    pretty = "Nearshore slopes online (m/m)"),

  c(col_name = "dist_deepest",
    description = "distance from the lake shoreline to the deepest point of the lake in meters",
    pretty = "Deepest point distance (m)"),

  c(col_name = "dists_deepest",
    description = "distance from the lake shoreline to the deepest point(s) of the lake in meters",
    pretty = "Mean deepest point distance (m)"),

  c(col_name = "dist_viscenter",
    description = "distance from the lake shoreline to the point furthest from the lake shoreline in meters",
    pretty = "Visual center distance (m)"),
  c(col_name = "reservoir_class", description = "classification of lakes are natural lakes or reservoirs by way of Polus et al.",
    pretty = "Reservoir class"),
  c(col_name = "shape_class",
    description = "classification of lakes as concave or convex following Hakanson (1977)",
    pretty = "Shape class"),
  c(col_name = "lake_perimeter_m",
    description = "perimeter of outer boundary of lake waterbody polygon from the NHD (excludes islands) by way of the LAGOS-US Locus module",
    pretty = "Perimeter (m)"),
  c(col_name = "lake_islandarea_ha",
    description = "surface area of islands within outer boundary of lake waterbody polygon from NHD by way of the LAGOS-US Locus module",
    pretty = "Island area (ha)"),
  c(col_name = "lake_elevation_m",
    description = "the elevation of the lake polygon centroid, in meters (referenced to the North American Vertical Datum of 1988 (NAVD88)) and obtained from the National Elevation Dataset by way of the LAGOS-US Locus module",
    pretty = "Elevation (m)"),
  c(col_name = "lake_shorelinedevfactor", description = "shoreline development factor calculated as the lake_perimeter_m / (2*sqrt(pi*lake_waterarea_ha*10000)) by way of the LAGOS-US Locus module",
    pretty = "Shoreline development"),
  c(col_name = "ws_lake_arearatio",
    description = "ratio between interlake watershed area and lake water area by way of the LAGOS-US Locus module",
    pretty = "Watershed-lake ratio"),
  c(col_name = "hu4_zoneid",
    description = "unique identifier assigned by LAGOS-US for zones in the spatial division HU4",
    pretty = "HUC4 ID"),
  c(col_name = "trig_depth",
    description = "Trigonometrically derived max depth (m)",
    pretty = "Geometric max depth (m)")
  )

write.csv(taxonomy_key, "data/taxonomy.csv", row.names = FALSE)
